OSX: Provide nanosecond resolution timer for tests
Directly implement the getCurrentTime function on OSX instead
of providing a gettimeofday based clock_gettime implementation;
we use mach_absolute_time together with mach_timebase_info.
NB: The multiplication is done first and then the division
because the risk of overflowing seems smaller then precision
loss together with the risk of actually having a ratio of zero,
i.e. when denom > numer, when doing it the other way around.