Removing the pedantic flag from gcc compiles
The library was not developed with the pedantic warning flag enabled, and
the build outputs a volume of verbose warning messages on every build.
It's not currently helpful to have this enabled.  This flag should be enabled
again as the warnings get fixed, to enable a more robust library.

Various warnings and #pragmas fixed that were remaining.
Changed the allocation of a temp buffer in corr-trmv.cpp