Output constants at the correct precision
Outputting a double-precision constant in a single-precision kernel
can make compilation impossible on very strict platforms if the device
has no double-precision capability. (For example, it causes an assert in
Beignet on Haswell).

Replace all 'naked' 0.0, 1.0 and 2.0 occurrences with the corresponding
integer and trust on automatic promotion.

For xgemm and its Python generator, replace the switch with a unified
definition using the DATA_TYPE_STRING.